update main_tab using GETX
