+++ client/lib/pages/main_tab/bindings/main_tab_binding.dart
@@ -1,0 +1,10 @@
+import 'package:get/get.dart';
+
+import '../controllers/main_tab_controller.dart';
+
+class MainTabBinding extends Bindings {
+  @override
+  void dependencies() {
+    Get.lazyPut<MainTabController>(() => MainTabController());
+  }
+}
