+++ client/lib/routes/app_pages.dart
@@ -1,3 +1,4 @@
+import 'package:client/pages/main_tab/bindings/main_tab_binding.dart';
 import 'package:get/get.dart';
 
 import '../pages/home_page/new_arrival/bindings/new_arrival_binding.dart';
@@ -15,7 +16,7 @@ import '../pages/home_page/home_screen/views/home_screen_view.dart';
 import '../pages/home_page/new_arrival/views/new_arrival_view.dart';
 import '../pages/login/bindings/login_binding.dart';
 import '../pages/login/views/login_view.dart';
-import '../pages/main_tab/main_tab.dart';
+import '../pages/main_tab/views/main_tab_view..dart';
 import '../pages/profile/change_password/bindings/change_password_binding.dart';
 import '../pages/profile/change_password/views/change_password_view.dart';
 import '../pages/profile/order_ongoing/bindings/order_ongoing_binding.dart';
@@ -95,6 +96,7 @@ class AppPages {
     GetPage(
       name: mainTab,
       page: () => const MainTab(),
+      binding: MainTabBinding(),
     ),
     GetPage(
       name: home,
@@ -160,4 +162,3 @@ class AppPages {
     ),
   ];
 }
-
